\name{fjolstTranslate-package}
\alias{fjolstTranslate-package}
\alias{fjolstTranslate}
\docType{package}
\title{
\packageTitle{fjolstTranslate}
}
\description{
\packageDescription{fjolstTranslate}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{fjolstTranslate}
\packageIndices{fjolstTranslate}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{fjolstTranslate}

Maintainer: \packageMaintainer{fjolstTranslate}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
